Move UserIri value object
